+++ server/service/ParsedownExtension.php
@@ -7,6 +7,53 @@ require_once __DIR__ . "/ext/Parsedown.php";
 class ParsedownExtension extends Parsedown
 {
     /**
+     * The private attribute to access the instance of the service class
+     * UserInput.
+     */
+    private $user_input;
+
+    /**
+     * The constructor to add the new inline parser for user form fields.
+     *
+     * @param instance $user_input
+     *  An instance of the service class UserInput.
+     */
+    function __construct($user_input = null)
+    {
+        $this->user_input = $user_input;
+        $this->InlineTypes['@'][]= 'UserFormField';
+        $this->inlineMarkerList .= '@';
+    }
+
+    /**
+     * Extend the parsedown parser with user form fields. Using the syntax
+     * defined in UserInput::get_input_value_pattern() user form fields can be
+     * placed in markdown and, thus, allow to format user input fileds in any
+     * way imaginable (as HTML is allowed in markdown).
+     */
+    protected function inlineUserFormField($excerpt)
+    {
+        if($this->user_input === null)
+            return null;
+        $pattern = '/' . $this->user_input->get_input_value_pattern() . '/';
+        if (preg_match($pattern, $excerpt['text'], $matches))
+        {
+            $value = $this->user_input->get_input_value_by_pattern($matches[0],
+                $_SESSION['id_user']);
+            return array(
+
+                // How many characters to advance the Parsedown's
+                // cursor after being done processing this tag.
+                'extent' => strlen($matches[0]),
+                'element' => array(
+                    'name' => 'span',
+                    'text' => $value ?? "Bad Syntax",
+                ),
+            );
+        }
+    }
+
+    /**
      * Extende the inlineImage parser by changing the base path of image
      * sources. Further, allow to specify the image width and height as well
      * as css classes that will be attached to the image. By default the classes